Conversation workload calculation improved. reassign cronjob implemented. Some minor improvement made.

+++ src/main/java/ytuce/gp/mfmsp/Pojo/RepresentativePojo.java
@@ -15,7 +15,7 @@ public class RepresentativePojo {
     private List<ConversationPojo> conversationList;
     private Integer successRate;
 
-    private Integer workload;
+    private Long workload;
 
     public static RepresentativePojo entityToPojoBuilder(Representative representative){
         if(representative==null){
